// src/app/shared/models/skill/cp2020-martial-art.ts
import { SourceBook } from './../sourcebook';
export interface Cp2020MartialArt {
  name: string;
  source: SourceBook;
  ipMod: number;
  strike: number;
  punch: number;
  kick: number;
  disarm: number;
  sweep: number;
  block: number;
  dodge: number;
  grapple: number;
  throw: number;
  hold: number;
  choke: number;
  escape: number;
  ram: number;
}